Individual Pipeline page
designed for the user to view information specific to a single pipeline while being able to interact with overall pipeline activity and personal feed activity in the second container below the tree graph. There is a CTA in the header that will bring the user to the chRIS store and view the pipeline there.